+++ platform/qt/src/http_request.cpp
@@ -31,7 +31,7 @@ HTTPRequest::~HTTPRequest()
 
 QUrl HTTPRequest::requestUrl() const
 {
-    return QUrl::fromPercentEncoding(QByteArray(m_resource.url.data(), m_resource.url.size()));
+    return QUrl::fromPercentEncoding(QByteArray(m_resource.url.data(), static_cast<int>(m_resource.url.size())));
 }
 
 QNetworkRequest HTTPRequest::networkRequest() const
@@ -44,7 +44,7 @@ QNetworkRequest HTTPRequest::networkRequest() const
 
     if (m_resource.priorEtag) {
         const auto etag = m_resource.priorEtag;
-        req.setRawHeader("If-None-Match", QByteArray(etag->data(), etag->size()));
+        req.setRawHeader("If-None-Match", QByteArray(etag->data(), static_cast<int>(etag->size())));
     } else if (m_resource.priorModified) {
         req.setRawHeader("If-Modified-Since", util::rfc1123(*m_resource.priorModified).c_str());
     }
